add notification when user turn off the syncing function
