+++ scripts/ui/config.js
@@ -1,3 +1,4 @@
+import { showToast } from "./showmessage.js";
 export function initConfigButton() {
   const btn = document.getElementById('floating-config-btn');
   if (!btn) return;
@@ -127,31 +128,8 @@ export function initSaveConfigButton(){
    if(saobangSelected) localStorage.setItem('saobang', saobangSelected.value);
    if(isAutoLichThiSelected) localStorage.setItem('autolichthi', isAutoLichThiSelected.value);
 
-   showConfigToast("Đã lưu cài đặt thành công!");
+   showToast("Đã lưu cài đặt thành công!");
   });
 
-  function showConfigToast(message) {
-    let toast = document.createElement('div');
-    toast.textContent = message;
-    toast.style.position = 'fixed';
-    toast.style.left = '50%';
-    toast.style.top = '50%';
-    toast.style.transform = 'translate(-50%, -50%)';
-    toast.style.background = 'linear-gradient(135deg,#21c7a8,#7c4dff)';
-    toast.style.color = '#fff';
-    toast.style.padding = '12px 28px';
-    toast.style.borderRadius = '24px';
-    toast.style.fontWeight = '600';
-    toast.style.fontSize = '1rem';
-    toast.style.boxShadow = '0 2px 12px rgba(33,199,168,0.13)';
-    toast.style.zIndex = '99999';
-    toast.style.opacity = '0';
-    toast.style.transition = 'opacity 0.3s';
-    document.body.appendChild(toast);
-    setTimeout(() => { toast.style.opacity = '1'; }, 10);
-    setTimeout(() => {
-      toast.style.opacity = '0';
-      setTimeout(() => { toast.remove(); }, 350);
-    }, 2000);
-  }
+  
 }
